Moved key pair test from ConfigTest to CryptoTest
